+++ tt_metal/impl/profiler/tt_metal_profiler.cpp
@@ -175,7 +175,7 @@ void syncDeviceHost(distributed::MeshDevice* mesh_device, IDevice* device, CoreC
 
     log_info(tt::LogMetal, "SYNC PROGRAM FINISH IS DONE ON {}", device_id);
     if ((profiler_state_manager->smallest_host_time.at(device_id) == 0) ||
-        (profiler_state_manager->smallest_host_time.at(device_id) > hostStartTime)) {
+        (profiler_state_manager->smallest_host_time.at(device_id) > static_cast<uint64_t>(hostStartTime))) {
         profiler_state_manager->smallest_host_time.at(device_id) = hostStartTime;
     }
 
@@ -253,8 +253,8 @@ void syncDeviceHost(distributed::MeshDevice* mesh_device, IDevice* device, CoreC
         log_file.open(log_path, std::ios_base::app);
     }
 
-    int init = profiler_state_manager->device_host_time_pair.at(device_id).size() - sampleCount;
-    for (int i = init; i < profiler_state_manager->device_host_time_pair.at(device_id).size(); i++) {
+    int init = static_cast<int>(profiler_state_manager->device_host_time_pair.at(device_id).size()) - sampleCount;
+    for (int i = init; i < static_cast<int>(profiler_state_manager->device_host_time_pair.at(device_id).size()); i++) {
         log_file << fmt::format(
                         "{:5},{:5},{:5},{:20},{:20},{:20.2f},{:20},{:20},{:20.2f},{:20.15f},{:20.15f},{:20},1.0,0",
                         device_id,
@@ -515,7 +515,7 @@ void syncAllDevices(ChipId host_connected_device) {
     for (auto& sender : profiler_state_manager->device_device_time_pair) {
         for (auto& receiver : sender.second) {
             std::vector<std::pair<uint64_t, uint64_t>> timePairs;
-            for (int i = 0; i < receiver.second.size(); i += 2) {
+            for (size_t i = 0; i < receiver.second.size(); i += 2) {
                 uint64_t senderTime = (receiver.second[i].first + receiver.second[i + 1].first) / 2;
                 timePairs.push_back({senderTime, receiver.second[i].second});
             }
